server2/requirements.txt
# GIFT ZONA Betting Platform - Python Dependencies
# Рефакторинг архитектуры с улучшенной структурой

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
asyncpg>=0.29.0

# Telegram
pyrogram>=2.0.0
pyrofork
tgcrypto

# HTTP Clients
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Environment & Config
python-dotenv>=1.0.0

# Image Processing
pillow>=10.0.0

# Scheduling & Background Tasks
apscheduler>=3.10.0

# Parsing & Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
cloudscraper>=1.2.0
curl_cffi>=0.5.0

# External APIs
portalsmp

# System & Process Management
psutil>=5.9.0

# WebSockets
websockets>=12.0

# Utilities
pydantic>=2.4.0

# Development (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0

